Progress bar works now.
It includes startdate anda enddate.

Since date and timestamps are both stored in the same column in both the databases, this resulted in ""datetime(date) BETWEEN datetime(?) AND datetime(?)"" not working correctly.

That was replaced with "'%d/%m/%Y  %H:%M:%S', date" and now it correctly fetch data for total seconds and hours between those dates. This is the only way it works.